+++ src/pages/Endpoint/Scan.vue
@@ -77,7 +77,6 @@
 
 <script lang="ts">
 import { defineComponent, onMounted, ref } from 'vue'
-import { useRoute } from 'vue-router'
 import { MainApi } from 'components/axios'
 import module from 'components/Module.vue'
 import crudBtn from 'components/Buttons/CrudBtn.vue'
@@ -137,8 +136,7 @@ export default defineComponent({
     scanInfo,
   },
   setup() {
-    const route = useRoute()
-    const scan = JSON.parse(route.params.scan as string) as Scan
+    const scan = JSON.parse(sessionStorage.getItem('scan') as string) as Scan
 
     return {
       scan,