+++ src/pages/Port/Scan.vue
@@ -1,6 +1,7 @@
 <template>
   <q-page class="q-pa-lg">
     <scan-info :scan="scan" />
+    <chart :id="scan.id" type="ports" a-name="Port" b-name="Protocol" />
 
     <module parent="Scan" icon="find_in_page" name="Ports">
       <template v-slot:card>
@@ -83,12 +84,12 @@
 
 <script lang="ts">
 import { defineComponent, onMounted, ref } from 'vue'
-import { useRoute } from 'vue-router'
 import { MainApi } from 'components/axios'
 import module from 'components/Module.vue'
 import crudBtn from 'components/Buttons/CrudBtn.vue'
 import actionBtn from 'components/Buttons/ActionBtn.vue'
 import scanInfo from 'components/ScanInfo.vue'
+import chart from 'components/Chart.vue'
 import { Port, col } from 'src/models/port'
 import { Scan } from 'src/models/scan'
 import { del, del_all, protocol2color, copy } from 'components/utils'
@@ -128,11 +129,18 @@ export default defineComponent({
     crudBtn,
     actionBtn,
     scanInfo,
+    chart,
   },
   setup() {
-    const route = useRoute()
-    const scan = JSON.parse(route.params.scan as string) as Scan
-    return { scan, ...useTable(scan.id), del, del_all, protocol2color, copy }
+    const scan = JSON.parse(sessionStorage.getItem('scan') as string) as Scan
+    return {
+      scan,
+      ...useTable(scan.id),
+      del,
+      del_all,
+      protocol2color,
+      copy,
+    }
   },
 })
 </script>